BDD/RobotFramework/requirements.txt
robotframework-seleniumlibrary==5.1.3
robotframework-databaselibrary==1.2.4
robotframework-requests==0.9.1
robotframework-datadriver==1.4.1
robotframework-pabot==2.0.1
allure-robotframework==2.9.43
PyMySQL==1.0.2